+++ test/pycardano/util.py
@@ -44,6 +44,7 @@ class FixedChainContext(ChainContext):
         collateral_percent=150,
         max_collateral_inputs=3,
         coins_per_utxo_word=34482,
+        coins_per_utxo_byte=4310,
     )
 
     _genesis_param = GenesisParameters(